test(ui): move code-block rendering tests beside their clipboard suite
ui/src/components/markdown.test.ts sits one line under the 700-line max-lines
cap after #141466, #141485 and #141497 each moved a slice out. Move the
toSanitizedMarkdownHtml code-block rendering cases into the existing
markdown-code-blocks.test.ts, share htmlFragment and withControlUiBasePath
from markdown.test-support.ts, and drop the copy of htmlFragment that
markdown-streaming.test.ts carried. No assertion changes; the case count across
ui/src/components/markdown*.test.ts is unchanged (232).
